refactor: BuildSearchIndexCommand and options
